feat: extracting process from led (ultimately making the latter into a sort of driver)
