+++ led_control_project/src/process.h
@@ -1,0 +1,8 @@
+#ifndef PROCESS_H
+#define PROCESS_H
+
+#include "silly_led_control.h"
+
+void silly_process(long period);
+
+#endif // PROCESS_H
