+++ led_control_project/src/silly_led_control.h
@@ -18,10 +18,6 @@
 #define GPIO_LED      "/sys/class/gpio/gpio10"
 #define LED           "10"
 
-/**
-*   THIS FUNCTION DOES NOT RETURN
-*/
-void led_process(long period);
 
 int open_led(void);
